Minor optimizer bug fix.
Projection of search direction into feasible directions had a division by magnitude rather
than correct division by squared magnitude. Additionally it was found that the solver had a
hard time with the inequality constraints for non-affine constraints with just a projection.
The solution was to add in an affine decorator for affine inequalities as well as adjust the
feasible direction slightly to incorporate a negative inequality direction.